+++ src/etc/index.ts
@@ -1,0 +1,3 @@
+export * from './could-be-type';
+export * from './get-type-services';
+export * from './is';