+++ src/index.ts
@@ -1,1 +1,15 @@
-module.exports.rules = require("requireindex")(`${__dirname}/rules`);
+import { name, version } from '../package.json';
+import { preferAsyncPipeRule } from './rules/prefer-async-pipe';
+import { preferCompositionRule } from './rules/prefer-composition';
+import { preferTakeuntilRule } from './rules/prefer-takeuntil';
+
+const rxjsAngularX = {
+  meta: { name, version },
+  rules: {
+    'prefer-async-pipe': preferAsyncPipeRule,
+    'prefer-composition': preferCompositionRule,
+    'prefer-takeuntil': preferTakeuntilRule,
+  },
+};
+
+export default rxjsAngularX;