+++ tests/utils.ts
@@ -1,3 +1,0 @@
-import { createRuleTester } from "eslint-etc";
-
-export const ruleTester = createRuleTester({});